requirements.txt
accelerate==0.34.2
contourpy==1.3.0
datasets==3.0.0
encodec==0.1.1
et-xmlfile==1.1.0
evaluate==0.4.3
huggingface-hub==0.25.0
IProgress==0.4
ipywidgets==8.1.5
librosa==0.10.2
numpy==2.0
openai-clip==1.0.1
opencv-python==4.10.0.84
openpyxl==3.1.5
pandas==2.2.2
pyarrow==17.0.0
python-dateutil==2.9.0
pytz==2024.2
seaborn==0.13.2
sentencepiece==0.2.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
transformers==4.44.2
torchvision==0.19.1+cu124
tzdata==2024.1
